requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-gcharts==1.3
django-google-charts==0.1.1
django-postgrespool==0.3.0
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6
python-dateutil==2.4.2
six==1.9.0
SQLAlchemy==1.0.0
static3==0.5.1
whitenoise==1.0.6
